Reorg: simpleStore, etcdStore
