--- examples/raft-on-mesh/store_test.go
@@ -1,11 +1,0 @@
-package main
-
-import "testing"
-
-func TestStateMachineSetGet(t *testing.T) {
-	t.Skip("TODO")
-}
-
-func TestStateMachineWatch(t *testing.T) {
-	t.Skip("TODO")
-}
